requirements.txt
tensorflow-cpu==2.8.1
numpy==1.21.5
flask==2.1.2
flask_cors==3.0.10
gunicorn==20.1.0